Consolidate counters to one file
