--- lib/sucker_punch/counter/busy.rb
@@ -1,23 +1,0 @@
-require 'sucker_punch/counter/utilities'
-
-module SuckerPunch
-  module Counter
-    class Busy
-      attr_reader :counter
-
-      include SuckerPunch::Counter::Utilities
-
-      COUNTER = Concurrent::Map.new do |hash, name| #:nodoc:
-        hash.compute_if_absent(name) { Concurrent::AtomicFixnum.new }
-      end
-
-      def self.clear
-        COUNTER.clear
-      end
-
-      def initialize(queue_name)
-        @counter = COUNTER[queue_name]
-      end
-    end
-  end
-end
