--- lib/sucker_punch/counter/utilities.rb
@@ -1,18 +1,0 @@
-module SuckerPunch
-  module Counter
-    module Utilities
-      def value
-        @counter.value
-      end
-
-      def increment
-        @counter.increment
-      end
-
-      def decrement
-        @counter.decrement
-      end
-    end
-  end
-end
-
